Switch to header based API authentication
Previously we relied on cookie based authentication. This is
problematic, because Next.js' `getInitialProps` is called on both
the server and the client. When it is called on the server, cookies
must be read from the request, rather than the browser.

This change allows `getInitialProps` to grab the user cookie regardless
of whether it is called on the server or the client. It then passes this
token to the API via the `Authorization` header, which is a more typical
way to authenticate APIs than using cookies.